# DataBase/apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: DataBase Gateway
  version: 0.1.0
  description: Thin API facade for the DataBase ecosystem.
servers:
  - url: http://127.0.0.1:18090
components:
  securitySchemes:
    DataBaseApiKey:
      type: apiKey
      in: header
      name: X-DataBase-Api-Key
      description: Optional. Required only when DATABASE_GATEWAY_AUTH_REQUIRED=true.
  schemas:
    RequestId:
      type: string
      description: Per-request correlation id returned in X-Request-Id.
...
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - message
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
        message:
          type: string
        requestId:
          $ref: "#/components/schemas/RequestId"
...
    MyBlogReaderMemory:
      type: object
      properties:
        id:
          type: string
        objectId:
          type: string
        objectType:
          type: string
        title:
          type: string
        href:
          type: string
        progress:
          type: number
        location: {}
        scrollTop:
          type: number
        timestamp:
          type: number
        lastReadAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - objectId
        - objectType
        - title
        - href
        - progress
        - location
        - scrollTop
        - timestamp
        - lastReadAt
        - updatedAt
      additionalProperties: false
    MyBlogReaderMemoryResponse:
      type: object
      properties:
        ok:
          type: boolean
        item:
          anyOf:
            - $ref: "#/components/schemas/MyBlogReaderMemory"
            - type: "null"
        requestId:
          type: string
      required:
        - ok
        - item
        - requestId
      additionalProperties: false
...
    MyBlogReaderMemoryQueryResponse:
      type: object
      properties:
        ok:
          type: boolean
        item:
          anyOf:
            - $ref: "#/components/schemas/MyBlogReaderMemory"
            - type: "null"
        items:
          type: array
          items:
            $ref: "#/components/schemas/MyBlogReaderMemory"
        requestId:
          type: string
      required:
        - ok
        - requestId
      additionalProperties: false
    MyBlogReaderMemoryUpsertRequest:
      type: object
      properties:
        id:
          type: string
        objectId:
          type: string
        objectType:
          type: string
        type:
          type: string
        title:
          type: string
        href:
          type: string
        progress:
          type: number
        percent:
          type: number
        location: {}
        scrollTop:
          type: number
        lastReadAt:
          type: number
        updatedAt:
          type: number
      additionalProperties: {}
    MyBlogReaderHighlight:
      type: object
      properties:
        id:
          type: string
        articleId:
          type: string
        objectId:
          type: string
        objectType:
          type: string
        title:
          type: string
        text:
          type: string
        color:
          type: string
        note:
          type: string
        anchor: {}
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - articleId
        - objectId
        - objectType
        - title
        - text
        - color
        - note
        - anchor
        - createdAt
        - updatedAt
      additionalProperties: false
    MyBlogReaderHighlightsResponse:
      type: object
      properties:
        ok:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/MyBlogReaderHighlight"
        requestId:
          type: string
      required:
        - ok
        - items
        - requestId
      additionalProperties: false
    MyBlogReaderHighlightResponse:
      type: object
      properties:
        ok:
          type: boolean
        item:
          $ref: "#/components/schemas/MyBlogReaderHighlight"
        requestId:
          type: string
      required:
        - ok
        - item
        - requestId
      additionalProperties: false
    MyBlogReaderHighlightUpsertRequest:
      type: object
      properties:
        id:
          type: string
        articleId:
          type: string
        objectId:
          type: string
        objectType:
          type: string
        type:
          type: string
        title:
          type: string
        text:
          type: string
        color:
          type: string
        note:
          anyOf:
            - type: string
            - type: "null"
        anchor: {}
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - text
      additionalProperties: {}
    MyBlogVisualSource:
      type: object
      properties:
        id:
          type: string
        sourceType:
          type: string
        provider:
          type: string
        sourceUrl:
          type: string
        boardId:
          type: string
        providerConfig:
          type: object
          propertyNames:
            type: string
          additionalProperties: {}
        title:
          type: string
        collectionTitle:
          type: string
        partitionPattern: {}
        syncIntervalSeconds:
          type: number
        lastCursor:
          type: string
        lastSyncedAt:
          anyOf:
            - type: number
            - type: "null"
        syncStatus:
          type: string
        pinsSnapshotHash:
          type: string
        lastError:
          type: string
        pinCount:
          type: number
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - sourceType
        - provider
        - sourceUrl
        - boardId
        - providerConfig
        - title
        - collectionTitle
        - partitionPattern
        - syncIntervalSeconds
        - lastCursor
        - lastSyncedAt
        - syncStatus
        - pinsSnapshotHash
        - lastError
        - pinCount
        - createdAt
        - updatedAt
      additionalProperties: false
    MyBlogVisualPin:
      type: object
      properties:
        id:
          type: string
        pinId:
          type: string
        sourceId:
          type: string
        pinUrl:
          type: string
        imagePreviewUrl:
          type: string
        title:
          type: string
        description:
          type: string
        boardId:
          type: string
        positionIndex:
          type: number
        downloaded:
          type: boolean
        raw: {}
        firstSeenAt:
          type: number
        lastSeenAt:
          type: number
        deletedAt:
          anyOf:
            - type: number
            - type: "null"
      required:
        - id
        - pinId
        - sourceId
        - pinUrl
        - imagePreviewUrl
        - title
        - description
        - boardId
        - positionIndex
        - downloaded
        - raw
        - firstSeenAt
        - lastSeenAt
        - deletedAt
      additionalProperties: false
    MyBlogVisualSnapshotResponse:
      type: object
      properties:
        ok:
          type: boolean
        version:
          type: number
        mode:
          type: string
        downloaded:
          type: boolean
        generatedAt:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/MyBlogVisualSource"
        pinsBySource:
          type: object
          propertyNames:
            type: string
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/MyBlogVisualPin"
        requestId:
          type: string
      required:
        - ok
        - version
        - mode
        - downloaded
        - generatedAt
        - sources
        - pinsBySource
        - requestId
      additionalProperties: false
    MyBlogVisualSyncResultRequest:
      type: object
      properties:
        sourceId:
          type: string
        provider:
          type: string
        ok:
          type: boolean
        runId:
          type: string
        snapshotHash:
          type: string
        error:
          anyOf:
            - type: string
            - type: "null"
        pins:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              pinId:
                type: string
              pinUrl:
                type: string
              imagePreviewUrl:
                type: string
              image:
                type: string
              title:
                type: string
              description:
                type: string
              boardId:
                type: string
              positionIndex:
                type: number
              raw: {}
            additionalProperties: {}
      additionalProperties: {}
    MyBlogVisualSyncResultResponse:
      type: object
      properties:
        ok:
          type: boolean
        sourceId:
          type: string
        syncedItems:
          type: number
        snapshotHash:
          type: string
        requestId:
          type: string
      required:
        - ok
        - sourceId
        - syncedItems
        - snapshotHash
        - requestId
      additionalProperties: false
...
paths:
...
  /myblog/runtime/reader/memory:
    get:
      operationId: getMyBlogReaderMemory
      summary: Read MyBlog reader memory from DataBase
      description: Returns either one reader-memory item by objectId or the latest reader-memory list. MyBlog must call this Gateway route instead of connecting to MySQL.
      parameters:
        - name: objectId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
            maximum: 50
      responses:
        "200":
          description: MyBlog reader memory projection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogReaderMemoryQueryResponse"
        "401":
          description: Missing or invalid API key.
    post:
      operationId: upsertMyBlogReaderMemory
      summary: Upsert MyBlog reader memory into DataBase
      description: Persists MyBlog reader progress through the Gateway-owned MySQL tables.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MyBlogReaderMemoryUpsertRequest"
      responses:
        "200":
          description: Upserted reader-memory item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogReaderMemoryResponse"
        "400":
          description: Invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
  /myblog/runtime/reader/highlights:
    get:
      operationId: listMyBlogReaderHighlights
      summary: Read MyBlog reader highlights from DataBase
      description: Returns reader highlights from DataBase-owned storage. MyBlog must not query reader_highlights directly.
      parameters:
        - name: objectId
          in: query
          schema:
            type: string
        - name: articleId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 200
      responses:
        "200":
          description: MyBlog reader highlights.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogReaderHighlightsResponse"
        "401":
          description: Missing or invalid API key.
    post:
      operationId: upsertMyBlogReaderHighlight
      summary: Upsert a MyBlog reader highlight into DataBase
      description: Persists a reader highlight through the Gateway-owned MySQL tables.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MyBlogReaderHighlightUpsertRequest"
      responses:
        "200":
          description: Upserted reader highlight.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogReaderHighlightResponse"
        "400":
          description: Invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
  /myblog/runtime/visuals/snapshot:
    get:
      operationId: getMyBlogVisualSnapshot
      summary: Read MyBlog visual runtime snapshot from DataBase
      description: Returns the stored visual-source and pin snapshot from Gateway-owned MySQL tables.
      responses:
        "200":
          description: MyBlog visual runtime snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogVisualSnapshotResponse"
        "401":
          description: Missing or invalid API key.
  /myblog/runtime/visuals/sync-result:
    post:
      operationId: recordMyBlogVisualSyncResult
      summary: Record a MyBlog visual sync result into DataBase
      description: MyBlog may execute Pinterest/Apify sync, but the resulting pins and source status are stored by DataBase Gateway.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MyBlogVisualSyncResultRequest"
      responses:
        "200":
          description: Stored visual sync result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBlogVisualSyncResultResponse"
        "401":
          description: Missing or invalid API key.
...